Use measurement_uid instead of report_id for requests
